refactor(runtime): move the offload and progress seams to a dependency-neutral substrate (FND-316)
`storage/` depended upward on `execution/` for `run_in_thread` and
`current_progress_tracker`, while `execution/_temporal/activities.py` reached
back down into `storage/`. The cycle was real; eleven lazy imports across
`storage/`, `app/` and `execution/` moved the failure from import time to never,
each carrying its own copy of the same `# noqa: PLC0415 — circular:` explanation.

The mechanism was never "storage imports execution" on its own: importing any
`execution` submodule runs `execution/__init__`, which reaches `storage.ops`
through `_temporal` -> `app.base` -> `contracts` -> `credentials` ->
`common.utils`. So a module-scope import from `execution` inside `storage.ops`
raised `ImportError`, while the same import in `storage/transfer.py` worked —
purely because that file is not in `storage/__init__`'s eager chain.

Move both primitives into `application_sdk/_runtime/`, a bottom layer every
other layer may import at module scope. `SerializableEnum` moves down with them
(`_runtime/progress.py` needs it, and `contracts.base` is not reachable from
`storage/`) and is re-exported from `contracts.base`, so the public path and the
class identity are unchanged. `execution/heartbeat.py` splits along the seam it
already had: the controllers, the auto-heartbeat loop and the stall watchdog
stay, the offload primitives move out and are re-exported. `execution/progress.py`
becomes a full re-export.

The app-facing paths do not change: `application_sdk.execution.heartbeat` and
`application_sdk.execution.progress` stay the documented imports and hand back
the same objects, so a patch applied at either path affects both. `_runtime` is
underscore-private on purpose — substrate, not a new public surface.

Every lazy import on the storage <-> execution edge is gone, including the one in
`storage/chunked.py:_fetch_chunk` that was paying a `sys.modules` lookup per
range GET. Two remain and neither is about storage: `observability/observability.py`
(an observability-internal cycle through `logger_adaptor`, now named as such) and
`activities.py`'s `app.context` import (the separate app <-> execution cycle).

`tests/unit/runtime/test_layering.py` enforces the rule in a subprocess rather
than leaving it to be rediscovered: each `_runtime` module must pull in no layer
above it, `storage.ops` must still load standalone without `execution`, and the
facades must re-export the same objects.

Conformance P031/P036 exempted `execution/heartbeat.py` by path suffix as the
sanctioned home of the thread and process pools. The exemption follows the
implementation to `_runtime/offload.py`; `heartbeat.py` is no longer exempt
because it no longer constructs either pool, and new tests pin both halves.
Findings across the repo are unchanged from `main` (492 before, 492 after).

ADR-0019 records the decision, the rule, and the two alternatives rejected
(documenting the constraint in place; making `execution/__init__` lazy).
