+++ application_sdk/execution/progress_telemetry.py
@@ -41,7 +41,7 @@ import functools
 from collections.abc import Callable
 from typing import Any
 
-from application_sdk.execution.progress import (
+from application_sdk._runtime.progress import (
     DEFAULT_MAX_NO_PROGRESS_SECONDS,
     ClosedHold,
     ProgressWatchdogMode,
