+++ application_sdk/storage/rolling.py
@@ -74,9 +74,10 @@ from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
 from typing import Any, Generic, TypeVar
 
+from application_sdk._runtime.offload import run_in_thread
+from application_sdk._runtime.progress import current_progress_tracker
 from application_sdk.common.file_ops import SafeFileOps
 from application_sdk.contracts.types import FileReference
-from application_sdk.execution.progress import current_progress_tracker
 from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.storage.rolling_errors import (
     InvalidRollingFileWriterError,
@@ -129,15 +130,6 @@ async def _call_maybe_blocking(fn: Callable[..., Any], *args: Any) -> Any:
     """
     if inspect.iscoroutinefunction(fn):
         return await fn(*args)
-    # Imported lazily, and it has to be: this module is in
-    # `storage/__init__`'s eager chain, so importing
-    # `application_sdk.execution.heartbeat` at module scope runs
-    # `execution/__init__` -> Temporal activity utils -> `application_sdk.app`
-    # -> back to a still-initialising `contracts.base`, raising ImportError.
-    # The cycle is via `execution/__init__`, not a direct storage -> execution
-    # edge. See `storage/batch.py` for the full chain.
-    from application_sdk.execution.heartbeat import run_in_thread  # noqa: PLC0415
-
     return await run_in_thread(fn, *args)
