+++ docs/adr/index.md
@@ -22,3 +22,4 @@ ADRs capture the architectural decisions behind the Application SDK — what was
 | [ADR-0016](0016-multi-pool-worker-routing.md) | Multi-Pool Worker Routing (N pools per app, @task(pool=...) static affinity) |
 | [ADR-0017](0017-native-execution-isolation.md) | Native Execution Isolation (best-effort work in an isolated process; run_fault_isolated / run_best_effort, conformance rule P036) |
 | [ADR-0018](0018-progress-aware-heartbeat.md) | Progress-Aware Stall Watchdog and Duration-Backstop Timeouts (in-process no-progress watchdog; heartbeat_timeout unchanged; start_to_close = backstop) |
+| [ADR-0019](0019-runtime-substrate-layer.md) | A Dependency-Neutral Runtime Substrate (`application_sdk._runtime`: offload + progress seams importable at module scope from any layer; kills the storage ↔ execution cycle) |
