+++ packages/conformance/conformance/suite/rules/determinism.py
@@ -281,7 +281,7 @@ RULES: tuple[RuleDefinition, ...] = (
             "``run_in_executor(<some-executor>, ...)`` with any executor other than\n"
             "``None`` is not flagged — a call-site-owned ``ThreadPoolExecutor`` is\n"
             "not the shared-pool contention this rule targets.\n"
-            "``application_sdk/execution/heartbeat.py`` is exempt: that is where\n"
+            "``application_sdk/_runtime/offload.py`` is exempt: that is where\n"
             "``run_in_thread()``'s own dedicated-executor dispatch lives.\n"
             "\n"
             "Remediation is a restructure (swap in ``run_in_thread()``), so findings\n"
@@ -332,7 +332,7 @@ RULES: tuple[RuleDefinition, ...] = (
             "receiver is not statically resolvable and is not flagged;\n"
             "``ThreadPoolExecutor`` is a thread pool, out of scope here (thread\n"
             "offload onto the shared default executor is governed by P031).\n"
-            "``application_sdk/execution/heartbeat.py`` is exempt — that is where the\n"
+            "``application_sdk/_runtime/offload.py`` is exempt — that is where the\n"
             "seam's own pool lives.\n"
             "\n"
             "Remediation is a restructure (route through the seam), so findings route\n"
